Module vostok_dao::proposal

Structs

NewProposal
NewProposal is an input to create a new Proposal. + voting_start must be after the current block number. + voting_duration must be between Contract.min_duration and Contract.max_duration. + execute_before is a last block number when the proposal can be closed and executed. Must be after voting_start + voting_duration.

Proposal

ProposalOut
JSON compatible return type for Proposal.

Voter

Enums

Action
Action is a JSON compatible type for encodidng actions

ActionInt
Internal Action representation

Functions

assert_valid_account